refactor(smEntities): changed location of zombie command
