+++ smEntities/src/main/kotlin/me/ixhbinphoenix/smPl/smEntities/Main.kt
@@ -1,8 +1,7 @@
 package me.ixhbinphoenix.smPl.smEntities
 
 import me.ixhbinphoenix.smPl.smCore.commands.BaseCommand
-import me.ixhbinphoenix.smPl.smEntities.entities.Npc
-import me.ixhbinphoenix.smPl.smEntities.entities.hostiles.Zombie
+import me.ixhbinphoenix.smPl.smEntities.commands.Commands
 import me.ixhbinphoenix.smPl.smEntities.events.Events
 import org.bukkit.ChatColor
 import org.bukkit.plugin.java.JavaPlugin
@@ -11,10 +10,9 @@ class Main : JavaPlugin() {
     init{
         instance = this
     }
-    var npcList: MutableList<Npc> = emptyList<Npc>().toMutableList()
     override fun onEnable() {
         val commands: HashMap<String, BaseCommand> = hashMapOf(
-            "zombie" to Zombie()
+            "zombie" to Commands()
         )
         for(cmd in commands){
             getCommand(cmd.key)?.setExecutor(cmd.value)
